requirements.txt
absl-py==2.3.1
accelerate==1.10.1
aiofiles==24.1.0
altair==5.5.0
annotated-types==0.7.0
anthropic==0.68.0
anyio==4.10.0
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1733250440834/work
attrs==25.3.0
azure-common==1.1.28
azure-core==1.35.1
azure-search-documents==11.5.3
blinker==1.9.0
Brotli==1.1.0
brotlicffi @ file:///C:/b/abs_f4lixvzmxt/croot/brotlicffi_1736183286104/work
cachetools==6.2.0
certifi @ file:///C:/b/abs_2bz30lzohz/croot/certifi_1754570654565/work/certifi
cffi @ file:///C:/b/abs_29_b57if3f/croot/cffi_1736184144340/work
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.2.1
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1733218098505/work
comm @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_comm_1753453984/work
contourpy==1.3.3
cycler==0.12.1
debugpy @ file:///C:/miniconda3/conda-bld/debugpy_1757067165604/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1740384970518/work
distro==1.9.0
docstring_parser==0.17.0
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1746947292760/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1756729339227/work
fastapi==0.117.1
ffmpy==0.6.1
filelock @ file:///C:/b/abs_758187j281/croot/filelock_1744281404850/work
fonttools==4.60.0
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
gmpy2 @ file:///C:/b/abs_d8ki0o0h97/croot/gmpy2_1738085498525/work
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.183.0
google-auth==2.41.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
gradio==5.47.2
gradio_client==1.13.3
groovy==0.1.2
grpcio==1.75.1
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
huggingface-hub==0.34.4
idna @ file:///C:/b/abs_aad84bnnw5/croot/idna_1714398896795/work
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_importlib-metadata_1747934053/work
ipykernel @ file:///D:/bld/ipykernel_1754352877088/work
ipython @ file:///D:/bld/bld/rattler-build_ipython_1759151854/work
ipython_pygments_lexers @ file:///home/conda/feedstock_root/build_artifacts/ipython_pygments_lexers_1737123620466/work
isodate==0.7.2
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1733300866624/work
Jinja2 @ file:///C:/b/abs_920kup4e6u/croot/jinja2_1741711580669/work
jiter==0.11.0
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1733440914442/work
jupyter_core @ file:///C:/b/abs_7c6fipznw3/croot/jupyter_core_1751991383524/work
kiwisolver==1.4.9
markdown-it-py==4.0.0
MarkupSafe @ file:///C:/b/abs_a0ma7ge0jc/croot/markupsafe_1738584052792/work
matplotlib==3.10.6
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1733416936468/work
mdurl==0.1.2
mkl-service==2.4.0
mkl_fft @ file:///C:/Users/dev-admin/mkl/mkl_fft_1730823082242/work
mkl_random @ file:///C:/Users/dev-admin/mkl/mkl_random_1730822522280/work
mpmath @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/mpmath_1699484863771/work
narwhals==2.5.0
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1733325553580/work
networkx @ file:///C:/b/abs_6c02vx1ct4/croot/networkx_1756709289161/work
nltk==3.9.1
numpy @ file:///C:/b/abs_0123vcxhf8/croot/numpy_and_numpy_base_1725470331966/work/dist/numpy-2.0.1-cp312-cp312-win_amd64.whl#sha256=e3e9d7fcaa32e9aa86aa46a7191a47fc4c4b35942462bbbedd17c45ef934dccd
openai==1.109.1
orjson==3.11.3
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1745345660/work
pandas==2.3.2
parso @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_parso_1755974222/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1733327343728/work
pillow @ file:///C:/b/abs_b50vowcrzo/croot/pillow_1738010273782/work
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_platformdirs_1756227402/work
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1756321756983/work
proto-plus==1.26.1
protobuf==5.29.5
psutil @ file:///C:/miniconda3/conda-bld/psutil_1757104556644/work
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1733569405015/work
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser @ file:///C:/miniconda3/conda-bld/pycparser_1757496054990/work
pydantic==2.11.9
pydantic_core==2.33.2
pydeck==0.9.1
pydub==0.25.1
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1750615794071/work
PyMuPDF==1.26.4
pyparsing==3.2.4
pyphen==0.17.2
PySocks @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/pysocks_1699473336188/work
pytesseract==0.3.13
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_python-dateutil_1751104122/work
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
pywin32 @ file:///C:/b/abs_6c9w_vp9zi/croot/pywin32_1754670534373/work
PyYAML @ file:///C:/b/abs_14xkfs39bx/croot/pyyaml_1728657968772/work
pyzmq @ file:///C:/miniconda3/conda-bld/pyzmq_1758103948070/work
referencing==0.36.2
regex==2025.9.1
requests @ file:///C:/b/abs_f53smfvswb/croot/requests_1756709376206/work
rich==14.1.0
rouge_score==0.1.2
rpds-py==0.27.1
rsa==4.9.1
ruff==0.13.2
sacremoses==0.1.1
safehttpx==0.1.6
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
semantic-version==2.10.0
sentencepiece==0.2.1
setuptools==78.1.1
shellingham==1.5.4
six @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_six_1753199211/work
smmap==5.0.2
sniffio==1.3.1
stack_data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1733569443808/work
starlette==0.48.0
streamlit==1.49.1
sympy==1.13.1
tenacity==9.1.2
textstat==0.7.10
threadpoolctl==3.6.0
tokenizers==0.22.0
toml==0.10.2
tomlkit==0.13.3
torch==2.6.0.dev20241112+cu121
torchaudio==2.5.0.dev20241112+cu121
torchvision==0.20.0.dev20241112+cu121
tornado @ file:///C:/b/abs_7dzpc171lf/croot/tornado_1748956950306/work
tqdm==4.67.1
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1733367359838/work
transformers==4.56.1
typer==0.19.2
typing-inspection==0.4.1
typing_extensions @ file:///C:/b/abs_ecq8gc0vbm/croot/typing_extensions_1756281142218/work
tzdata==2025.2
uritemplate==4.2.0
urllib3 @ file:///C:/b/abs_393g1ayee_/croot/urllib3_1750775481226/work
uvicorn==0.37.0
watchdog==6.0.0
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1758622279606/work
websockets==15.0.1
wheel==0.45.1
win-inet-pton @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/win_inet_pton_1699472992992/work
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1749421620841/work
